some refactor done, added Tracker class
